🎨 refactor code and abstract generateQuestionFromDB and generateQuestionFromGPT into seperate file
